add minor changes/fixes

+++ handlers/tile.go
@@ -11,7 +11,7 @@ const (
 	earthRadius              = 6378137.0
 	earthCircumference       = math.Pi * earthRadius
 	initialResolution        = 2 * earthCircumference / 256
-	dpi                uint8 = 96
+	dpi                int   = 96
 )
 
 type tileCoord struct {
@@ -59,7 +59,7 @@ func tileCoordFromString(z, x, y string) (tc tileCoord, ext string, err error) {
 	return
 }
 
-func calcScaleResolution(zoomLevel int, dpi uint8) (float64, float64) {
+func calcScaleResolution(zoomLevel int, dpi int) (float64, float64) {
 	var denom = 1 << zoomLevel
 	resolution := initialResolution / float64(denom)
 	scale := float64(dpi) * 39.37 * resolution // 39.37 in/m
